New Record 95

+++ SnakeAI/core/src/com/snake/ai/SavedSnakes.java
@@ -315,7 +315,7 @@ public class SavedSnakes implements Screen {
         }
         prefs.putFloat("gameNr " + gameNr + "average Fitness", maxFitness / POPULATIONSIZE);
         prefs.putInteger("gameNr " + gameNr + "hiScore", hiScore);
-        prefs.putInteger("gameNr " + gameNr + "bestSnakeEver.fitness", bestSnakeEver.bestSnakeEver.fitness);
+        prefs.putInteger("gameNr " + gameNr + "bestSnakeEver.fitness", (int) bestSnakeEver.bestSnakeEver.fitness);
         prefs.putInteger("gameNr " + gameNr + "population", population);
 
         //Neuronales Netzwerk Eigenschaften
